Move PacketGetPlayers and PacketPlayersReady to .network
These are used by the environmental controller too, so they shouldn't live
in the .blocks.teleporter package.
